# Speedup proof verification using batched inverse.